+++ changelog.md
@@ -92,6 +92,7 @@
 - PayPal credit card: Removed address fields and integrated Strong Customer Authentication (SCA) instead.
 - #834 Make the expiration date of visitor cookies of guests configurable.
 - #493 Display quantity name plural when quantity unit is more then 1.
+- #847 EU check VAT service: switch from SOAP to REST API 
 
 ### Bugfixes